requirements.txt
python-telegram-bot==21.5
docker==7.1.0
